+++ raisim/m1/include/raisim/object/ArticulatedSystem/ArticulatedSystem.hpp
@@ -23,6 +23,10 @@
 #include "raisim/object/singleBodies/Mesh.hpp"
 #include "raisim/contact/BisectionContactSolver.hpp"
 #include "raisim/constraints/PinConstraint.hpp"
+#include "raisim/sensors/RGBSensor.hpp"
+#include "raisim/sensors/DepthSensor.hpp"
+#include "raisim/sensors/InertialMeasurementUnit.hpp"
+
 
 namespace raisim {
 
@@ -131,7 +135,7 @@ class ArticulatedSystem : public Object {
       system_->updateMassInfo();
     }
 
-    double getWeight() const {
+    [[nodiscard]] double getWeight() const {
       return system_->getMass(localId_);
     }
 
@@ -139,7 +143,7 @@ class ArticulatedSystem : public Object {
       system_->getInertia()[localId_] = inertia;
     }
 
-    const Mat<3, 3> &getInertia() const {
+    [[nodiscard]] const Mat<3, 3> &getInertia() const {
       return system_->getInertia()[localId_];
     }
 
@@ -149,15 +153,15 @@ class ArticulatedSystem : public Object {
     }
 
     /* get center of mass position in parent frame */
-    const Vec<3> &getComPositionInParentFrame() const {
+    [[nodiscard]] const Vec<3> &getComPositionInParentFrame() const {
       return system_->getBodyCOM_B()[localId_];
     }
 
-    const std::unordered_map<std::string, CollisionDefinition *> &getCollisionSet() const {
+    [[nodiscard]] const std::unordered_map<std::string, CollisionDefinition *> &getCollisionSet() const {
       return colDef_;
     }
 
-    const std::unordered_map<std::string, VisObject *> &getVisualSet() const {
+    [[nodiscard]] const std::unordered_map<std::string, VisObject *> &getVisualSet() const {
       return visDef_;
     }
 
@@ -192,7 +196,7 @@ class ArticulatedSystem : public Object {
       system_->getFramePosition(frameId_, position);
     }
 
-    Mat<3, 3> getOrientation() const {
+    [[nodiscard]] Mat<3, 3> getOrientation() const {
       Mat<3, 3> orientation;
       system_->getFrameOrientation(frameId_, orientation);
       return orientation;
@@ -216,29 +220,29 @@ class ArticulatedSystem : public Object {
       coordinate[0] = system_->getGeneralizedCoordinate()[gcIndx_];
     }
 
-    double getJointAngle() const {
+    [[nodiscard]] double getJointAngle() const {
       return system_->getGeneralizedCoordinate()[gcIndx_];
     }
 
-    const raisim::Vec<3> &getPositionInParentFrame() const {
+    [[nodiscard]] const raisim::Vec<3> &getPositionInParentFrame() const {
       RSFATAL_IF(!isMovable_, "This is a fixed joint. You cannot change the position of a fixed joint.")
       return system_->getJointPos_P()[gvIndx_];
     }
 
-    const raisim::Vec<3> &getJointAxis() const {
+    [[nodiscard]] const raisim::Vec<3> &getJointAxis() const {
       RSFATAL_IF(!isMovable_, "This is a fixed joint. You cannot change the position of a fixed joint.")
       return system_->getJointAxis_P()[gvIndx_];
     }
 
-    Joint::Type getType() const {
+    [[nodiscard]] Joint::Type getType() const {
       return system_->getJointType(gvIndx_);
     }
 
-    size_t getIdxInGeneralizedCoordinate() const {
+    [[nodiscard]] size_t getIdxInGeneralizedCoordinate() const {
       return gvIndx_;
     }
 
-    Vec<3> getLinearVelocity() const {
+    [[nodiscard]] Vec<3> getLinearVelocity() const {
       Vec<3> linVel;
       system_->getFrameVelocity(frameId_, linVel);
       return linVel;
@@ -255,6 +259,7 @@ class ArticulatedSystem : public Object {
   friend class raisim::World;
   friend class raisim::urdf::LoadFromURDF2;
   friend class raisim::mjcf::LoadFromMjcf;
+  friend class raisim::InertialMeasurementUnit;
 
  public:
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
@@ -281,11 +286,16 @@ class ArticulatedSystem : public Object {
 
   /**
    * @return generalized coordinate of the system */
-  const raisim::VecDyn &getGeneralizedCoordinate() const { return gc_; }
+  [[nodiscard]] const raisim::VecDyn &getGeneralizedCoordinate() const { return gc_; }
 
   /**
    * @return generalized velocity of the system */
-  const raisim::VecDyn &getGeneralizedVelocity() const { return gv_; }
+  [[nodiscard]] const raisim::VecDyn &getGeneralizedVelocity() const { return gv_; }
+
+  /**
+   * @return generalized acceleration of the system. The generalized acceleration is computed in the integration step.
+   * So this function does not work properly if you have not integrated the world. */
+  [[nodiscard]] const raisim::VecDyn &getGeneralizedAcceleration() const { return ga_; }
 
   /**
    * @param[out] quaternion orientation of base*/
@@ -297,7 +307,7 @@ class ArticulatedSystem : public Object {
 
   /**
    * @return orientation of base*/
-  const raisim::Mat<3, 3> &getBaseOrientation() const { return rot_WB[0]; }
+  [[nodiscard]] const raisim::Mat<3, 3> &getBaseOrientation() const { return rot_WB[0]; }
 
   /**
    * @param[out] position position of base*/
@@ -310,7 +320,7 @@ class ArticulatedSystem : public Object {
 
   /**
    * @return position of base*/
-  raisim::Vec<3> getBasePosition() const {
+  [[nodiscard]] raisim::Vec<3> getBasePosition() const {
     RSFATAL_IF(jointType[0] != Joint::Type::FLOATING, "This method is only for floating base")
     raisim::Vec<3> position;
     position[0] = gc_[0];
@@ -344,12 +354,12 @@ class ArticulatedSystem : public Object {
   /**
    * set the generalized velocity
    * @param[in] jointVel the generalized velocity*/
-  void setGeneralizedVelocity(const Eigen::VectorXd &jointVel) { gv_ = jointVel; }
+  void setGeneralizedVelocity(const Eigen::VectorXd &jointVel) { gv_ = jointVel; updateKinematics();}
 
   /**
    * set the generalized velocity
    * @param[in] jointVel the generalized velocity*/
-  void setGeneralizedVelocity(const raisim::VecDyn &jointVel) { gv_ = jointVel; }
+  void setGeneralizedVelocity(const raisim::VecDyn &jointVel) { gv_ = jointVel; updateKinematics();}
 
   /**
    * set the generalized coordinsate of each joint in order.
@@ -414,7 +424,7 @@ class ArticulatedSystem : public Object {
    * The PD controller is implicit (using a continuous, linear model) so we cannot get the true gen force.
    * But if you set the time step small enough, the difference is negligible.
    * @return the generalized force */
-  VecDyn getGeneralizedForce() const {
+  [[nodiscard]] VecDyn getGeneralizedForce() const {
     VecDyn genForce(dof);
     genForce = tauFF_;
     if (controlMode_ == ControlMode::PD_PLUS_FEEDFORWARD_TORQUE) {
@@ -422,43 +432,54 @@ class ArticulatedSystem : public Object {
       vecvecCwiseMulThenSub(kd_, gv_, genForce);
       vecvecCwiseMulThenAdd(kd_, uref_, genForce);
     }
+
+    /// enforce joint torque limits
+    for (size_t i = 0; i < dof; i++) {
+      genForce[i] = genForce[i] < tauUpper_[i] ? genForce[i] : tauUpper_[i];
+      genForce[i] = genForce[i] > tauLower_[i] ? genForce[i] : tauLower_[i];
+    }
     return genForce;
   }
 
   /**
    * get the feedfoward generalized force (which is set by the user)
    * @return the feedforward generalized force */
-  const VecDyn &getFeedForwardGeneralizedForce() const { return tauFF_; }
+  [[nodiscard]] const VecDyn &getFeedForwardGeneralizedForce() const { return tauFF_; }
 
   /**
    * get the mass matrix
    * @return the mass matrix. Check Object/ArticulatedSystem section in the manual */
-  const MatDyn &getMassMatrix() { computeMassMatrix(M_); return M_; }
+  [[nodiscard]] const MatDyn &getMassMatrix() { computeMassMatrix(M_); return M_; }
 
   /**
    * get the coriolis and the gravitational term
    * @param[in] gravity gravitational acceleration. You should get this value from the world.getGravity();
    * @return the coriolis and the gravitational term. Check Object/ArticulatedSystem section in the manual */
-  const VecDyn &getNonlinearities(const Vec<3>& gravity) { computeNonlinearities(gravity, h_); return h_; }
+  [[nodiscard]] const VecDyn &getNonlinearities(const Vec<3>& gravity) {
+    VecDyn ga(dof);
+    ga.setZero();
+    recursiveNewtonEuler(gravity, false, false, ga, h_);
+    return h_;
+  }
 
   /**
    * get the inverse mass matrix. Note that this is actually damped inverse.
    * It contains the effect of damping and the spring effects due to the implicit integration.
    * YOU MUST CALL getMassMatrix FIRST BEFORE CALLING THIS METHOD.
    * @return the inverse mass matrix. Check Object/ArticulatedSystem section in the manual */
-  const MatDyn &getInverseMassMatrix() { computeSparseInverse(M_, Minv_); return Minv_; }
+  [[nodiscard]] const MatDyn &getInverseMassMatrix() { computeSparseInverse(M_, Minv_); return Minv_; }
 
   /**
    * get the center of mass of a composite body containing body i and all its children.
    * if you want the COM of the whole robot, just take the first element
    * This only works if you have called getMassMatrix() with the current state
    * @return the center of mass of the composite body */
-  const std::vector<raisim::Vec<3>> &getCompositeCOM() const { return composite_com_W; }
+  [[nodiscard]] const std::vector<raisim::Vec<3>> &getCompositeCOM() const { return composite_com_W; }
 
   /**
    * get the center of mass of the whole system
    * @return the center of mass of the system */
-  Vec<3> getCOM() const {
+  [[nodiscard]] Vec<3> getCOM() const {
     Vec<3> com = {0,0,0};
     double massTotal = 0;
     for(size_t i=0; i<nbody; i++) {
@@ -473,17 +494,17 @@ class ArticulatedSystem : public Object {
   /**
    * get the current composite inertia of a composite body containing body i and all its children
    * @return the inertia of the composite system */
-  const std::vector<raisim::Mat<3, 3>> &getCompositeInertia() const { return compositeInertia_W; }
+  [[nodiscard]] const std::vector<raisim::Mat<3, 3>> &getCompositeInertia() const { return compositeInertia_W; }
 
   /**
    * get the current composite mass of a composite body containing body i and all its children
    * @return get the composite mass */
-  const std::vector<double> &getCompositeMass() const { return compositeMass; }
+  [[nodiscard]] const std::vector<double> &getCompositeMass() const { return compositeMass; }
 
   /**
    * linear momentum of the whole system
    * @return momentum */
-  Vec<3> getLinearMomentum() const {
+  [[nodiscard]] Vec<3> getLinearMomentum() const {
     Vec<3> linearMomentum;
     linearMomentum.setZero();
 
@@ -500,7 +521,7 @@ class ArticulatedSystem : public Object {
    * returns the generalized momentum which is M * u
    * It is already computed in "integrate1()" so you don't have to compute again.
    * @return the generalized momentum */
-  const VecDyn &getGeneralizedMomentum() const { return generalizedMomentum_; }
+  [[nodiscard]] const VecDyn &getGeneralizedMomentum() const { return generalizedMomentum_; }
 
   /**
    * @param[in] gravity gravitational acceleration
@@ -514,7 +535,7 @@ class ArticulatedSystem : public Object {
   /**
    * @param[in] gravity gravitational acceleration
    * @return the potential energy (relative to zero height) given the gravity vector */
-  double getPotentialEnergy(const Vec<3> &gravity) const;
+  [[nodiscard]] double getPotentialEnergy(const Vec<3> &gravity) const;
 
   /**
    * @param[in] referencePoint the reference point about which the angular momentum is computed
@@ -537,7 +558,7 @@ class ArticulatedSystem : public Object {
    * getMovableJointNames. Note! the order doesn't correspond to dof since there are
    * joints with multiple dof's
    * @return movable joint names in the joint order. */
-  const std::vector<std::string> &getMovableJointNames() const { return movableJointNames; };
+  [[nodiscard]] const std::vector<std::string> &getMovableJointNames() const { return movableJointNames; };
 
   /**
    * @param[in] bodyIdx The body which contains the point, can be retrieved by getBodyIdx()
@@ -551,7 +572,7 @@ class ArticulatedSystem : public Object {
    * @param[in] nm name of the frame
    * @return the coordinate frame of the given name */
   CoordinateFrame &getFrameByName(const std::string &nm) { return frameOfInterest_[getFrameIdxByName(nm)]; }
-  const CoordinateFrame &getFrameByName(const std::string &nm) const { return frameOfInterest_[getFrameIdxByName(nm)]; }
+  [[nodiscard]] const CoordinateFrame &getFrameByName(const std::string &nm) const { return frameOfInterest_[getFrameIdxByName(nm)]; }
 
   /**
    * Refer to Object/ArticulatedSystem/Kinematics/Frame in the manual for details
@@ -561,7 +582,7 @@ class ArticulatedSystem : public Object {
     return *std::find_if(frameOfInterest_.begin(), frameOfInterest_.end(),
                         [name](const raisim::CoordinateFrame &ref) { return ref.bodyName == name; });
   }
-  const CoordinateFrame &getFrameByLinkName(const std::string &name) const {
+  [[nodiscard]] const CoordinateFrame &getFrameByLinkName(const std::string &name) const {
     return *std::find_if(frameOfInterest_.begin(), frameOfInterest_.end(),
                          [name](const raisim::CoordinateFrame &ref) { return ref.bodyName == name; });
   }
@@ -570,7 +591,7 @@ class ArticulatedSystem : public Object {
    * Refer to Object/ArticulatedSystem/Kinematics/Frame in the manual for details
    * @param[in] name name of the urdf link that is a child of the joint
    * @return the coordinate frame index of the given link name */
-  size_t getFrameIdxByLinkName(const std::string &name) const {
+  [[nodiscard]] size_t getFrameIdxByLinkName(const std::string &name) const {
     return std::find_if(frameOfInterest_.begin(), frameOfInterest_.end(),
                         [name](const raisim::CoordinateFrame &ref) { return ref.bodyName == name; })
                         - frameOfInterest_.begin();
@@ -581,21 +602,21 @@ class ArticulatedSystem : public Object {
    * @param[in] idx index of the frame
    * @return the coordinate frame of the given index */
   CoordinateFrame &getFrameByIdx(size_t idx) { return frameOfInterest_[idx]; }
-  const CoordinateFrame &getFrameByIdx(size_t idx) const { return frameOfInterest_[idx]; }
+  [[nodiscard]] const CoordinateFrame &getFrameByIdx(size_t idx) const { return frameOfInterest_[idx]; }
 
   /**
    * Refer to Object/ArticulatedSystem/Kinematics/Frame in the manual for details
    * The frame can be retrieved as as->getFrames[index]. This way is more efficient than above methods that use the frame name
    * @param[in] nm name of the frame
    * @return the index of the coordinate frame of the given index. Returns size_t(-1) if it doesn't exist */
-  size_t getFrameIdxByName(const std::string &nm) const;
+  [[nodiscard]] size_t getFrameIdxByName(const std::string &nm) const;
 
   /**
    * Refer to Object/ArticulatedSystem/Kinematics/Frame in the manual for details
    * The frame can be retrieved as as->getFrames[index]. This way is more efficient than above methods that use the frame name
    * @return a vector of the coordinate frames */
   std::vector<CoordinateFrame> &getFrames() { return frameOfInterest_; };
-  const std::vector<CoordinateFrame> &getFrames() const { return frameOfInterest_; };
+  [[nodiscard]] const std::vector<CoordinateFrame> &getFrames() const { return frameOfInterest_; };
 
   /**
    * @param[in] frameId the frame id which can be obtained by getFrameIdxByName()
@@ -671,6 +692,42 @@ class ArticulatedSystem : public Object {
    * @param[in] frame custom frame defined by the user
    * @param[out] angVel_W the angular velocity of the frame expressed to the world frame */
   void getFrameAngularVelocity(const CoordinateFrame &frame, Vec<3> &angVel_W) const;
+
+  /**
+   * YOU NEED TO ENABLE INVERSEDYNAMICS COMPUTATION TO USE THIS METHOD
+   * This is time derivative of (body velocity expressed in the world frame).
+   * If you premultiply this by the transpose of the rotation matrix of the frame,
+   * it is the same as what your accelerometer sensor will measure if you place it at the frame
+   * (but an IMU has gravity bias so they will be off by the gravitational acceleration).
+   * This is not the same as {(time derivative of body velocity expressed in the body frame) expressed in the world frame}
+   * @param[in] frameName name of the frame
+   * @param[out] acc_W the linear acceleration of the frame expressed in the world frame */
+  void getFrameAcceleration(const std::string &frameName, Vec<3> &acc_W) const {
+    getFrameAcceleration(frameOfInterest_[getFrameIdxByName(frameName)], acc_W);
+  }
+
+  /**
+   * YOU NEED TO ENABLE INVERSEDYNAMICS COMPUTATION TO USE THIS METHOD
+   * This is time derivative of (body velocity expressed in the world frame).
+   * If you premultiply this by the transpose of the rotation matrix of the frame,
+   * it is the same as what your accelerometer sensor will measure if you place it at the frame.
+   * (but an IMU has gravity bias so they will be off by the gravitational acceleration).
+   * This is not the same as {(time derivative of body velocity expressed in the body frame) expressed in the world frame}
+   * @param[in] frame custom frame defined by the user
+   * @param[out] acc_W the linear acceleration of the frame expressed in the world frame */
+  void getFrameAcceleration(const CoordinateFrame &frame, Vec<3> &acc_W) const {
+    Vec<3> pos;
+    acc_W.setZero();
+
+    getFramePosition(frame, pos);
+    acc_W = bodyLinearAcc[frame.parentId];
+
+    /// Centrifugal acc
+    addCentrifugalTerm(bodyAngVel_W[frame.parentId], pos - jointPos_W[frame.parentId], acc_W);
+
+    /// Linear acc due to ang acc
+    crossThenAdd(bodyAngAcc[frame.parentId], pos - jointPos_W[frame.parentId], acc_W);
+  }
 
   /**
    * @param[in] bodyIdx the body index. Note that body index and the joint index are the same because every body has one parent joint. It can be retrieved by getBodyIdx()
@@ -813,7 +870,7 @@ class ArticulatedSystem : public Object {
    * @param[in] bodyIdx the body index. it can be retrieved by getBodyIdx()
    * @param[in] posInBodyFrame the position of the point of interest expressed in the body frame
    * @param[out] pointVel the velocity of the point expressed in the world frame */
-  void getVelocity(size_t bodyIdx, const Vec<3> &posInBodyFrame, Vec<3> &pointVel) const;
+  void getVelocity(size_t bodyIdx, const Vec<3> &posInBodyFrame, Vec<3> &pointVel) const final;
 
   /**
    * @param[in] bodyIdx the body index. it can be retrieved by getBodyIdx()
@@ -827,19 +884,19 @@ class ArticulatedSystem : public Object {
    * returns the index of the body
    * @param[in] nm name of the body. The body name is the name of the movable link of the body
    * @return the index of the body. Returns size_t(-1) if the body doesn't exist. */
-  size_t getBodyIdx(const std::string &nm) const;
+  [[nodiscard]] size_t getBodyIdx(const std::string &nm) const;
 
   /**
    * @return the degrees of freedom */
-  size_t getDOF() const;
+  [[nodiscard]] size_t getDOF() const;
 
   /**
    * @return the dimension of generalized velocity (do the same thing with getDOF) */
-  size_t getGeneralizedVelocityDim() const;
+  [[nodiscard]] size_t getGeneralizedVelocityDim() const;
 
   /**
    * @return the dimension of generalized coordinate */
-  size_t getGeneralizedCoordinateDim() const;
+  [[nodiscard]] size_t getGeneralizedCoordinateDim() const;
 
   /**
    * The body pose is the pose of its parent joint (after its joint transformation)
@@ -871,47 +928,47 @@ class ArticulatedSystem : public Object {
   /**
    * @return a reference to joint position relative to its parent, expressed in the parent frame. */
   std::vector<raisim::Vec<3>> &getJointPos_P() { return jointPos_P; }
-  const std::vector<raisim::Vec<3>> &getJointPos_P() const { return jointPos_P; }
+  [[nodiscard]] const std::vector<raisim::Vec<3>> &getJointPos_P() const { return jointPos_P; }
 
   /**
    * @return a reference to joint orientation relative to its parent. */
   std::vector<raisim::Mat<3, 3>> &getJointOrientation_P() { return rot_JB; }
-  const std::vector<raisim::Mat<3, 3>> &getJointOrientation_P() const { return rot_JB; }
+  [[nodiscard]] const std::vector<raisim::Mat<3, 3>> &getJointOrientation_P() const { return rot_JB; }
 
   /**
    * @return a reference to joint axis relative to its parent, expressed in the parent frame. */
   std::vector<raisim::Vec<3>> &getJointAxis_P() { return jointAxis_P; }
-  const std::vector<raisim::Vec<3>> &getJointAxis_P() const { return jointAxis_P; }
+  [[nodiscard]] const std::vector<raisim::Vec<3>> &getJointAxis_P() const { return jointAxis_P; }
 
   /**
    * @return a reference to joint axis expressed in the world frame. */
-  const raisim::Vec<3> &getJointAxis(size_t idx) const { return jointAxis_W[idx]; }
+  [[nodiscard]] const raisim::Vec<3> &getJointAxis(size_t idx) const { return jointAxis_W[idx]; }
 
   /**
    * You MUST call updateMassInfo() after you change the mass
    * @return a reference to mass of each body.*/
   std::vector<double> &getMass() { return mass; }
-  const std::vector<double> &getMass() const { return mass; }
+  [[nodiscard]] const std::vector<double> &getMass() const { return mass; }
 
   /**
    * @return a reference to inertia of each body.*/
   std::vector<raisim::Mat<3, 3>> &getInertia() { return inertia_comB; }
-  const std::vector<raisim::Mat<3, 3>> &getInertia() const { return inertia_comB; }
+  [[nodiscard]] const std::vector<raisim::Mat<3, 3>> &getInertia() const { return inertia_comB; }
 
   /**
    * @return a reference to the position of the center of the mass of each body in the body frame.*/
   std::vector<raisim::Vec<3>> &getBodyCOM_B() { return comPos_B; }
-  const std::vector<raisim::Vec<3>> &getBodyCOM_B() const { return comPos_B; }
+  [[nodiscard]] const std::vector<raisim::Vec<3>> &getBodyCOM_B() const { return comPos_B; }
 
   /**
- * @return a reference to the position of the center of the mass of each body in the world frame.*/
+   * @return a reference to the position of the center of the mass of each body in the world frame.*/
   std::vector<raisim::Vec<3>> &getBodyCOM_W() { return comPos_W; }
-  const std::vector<raisim::Vec<3>> &getBodyCOM_W() const { return comPos_W; }
+  [[nodiscard]] const std::vector<raisim::Vec<3>> &getBodyCOM_W() const { return comPos_W; }
 
   /**
    * @return a reference to the collision bodies. Position and orientation can be set dynamically */
   raisim::CollisionSet &getCollisionBodies() { return collisionBodies; }
-  const raisim::CollisionSet &getCollisionBodies() const { return collisionBodies; }
+  [[nodiscard]] const raisim::CollisionSet &getCollisionBodies() const { return collisionBodies; }
 
   /**
    * @param name collision body name which is "LINK_NAME" + "/" + "COLLISION_NUMBER". For example, the first collision body of the link "base" is named as "base/0"
@@ -920,7 +977,7 @@ class ArticulatedSystem : public Object {
     return *std::find_if(collisionBodies.begin(), collisionBodies.end(),
                          [name](const raisim::CollisionDefinition &ref) { return ref.colObj->name == name; });
   }
-  const raisim::CollisionDefinition &getCollisionBody(const std::string &name) const {
+  [[nodiscard]] const raisim::CollisionDefinition &getCollisionBody(const std::string &name) const {
     return *std::find_if(collisionBodies.begin(), collisionBodies.end(),
                          [name](const raisim::CollisionDefinition &ref) { return ref.colObj->name == name; });
   }
@@ -933,7 +990,7 @@ class ArticulatedSystem : public Object {
   /**
    * @param[in] bodyIdx the body index. it can be retrieved by getBodyIdx()
    * @return mass of the body */
-  double getMass(size_t bodyIdx) const final { return mass[bodyIdx]; }
+  [[nodiscard]] double getMass(size_t bodyIdx) const final { return mass[bodyIdx]; }
 
   /**
    * set body mass. It is indexed for each body, not for individual link. Check this link
@@ -943,7 +1000,7 @@ class ArticulatedSystem : public Object {
 
   /**
    * @return the total mass of the system.*/
-  double getTotalMass() const { return compositeMass[0]; }
+  [[nodiscard]] double getTotalMass() const { return compositeMass[0]; }
 
   /**
    * set external forces or torques expressed in the world frame acting on the COM of the body.
@@ -980,7 +1037,8 @@ class ArticulatedSystem : public Object {
    * @param[in] force the applied force in the world frame*/
   void setExternalForce(const std::string &frame_name, const Vec<3> &force) {
     auto &frame = getFrameByName(frame_name);
-    setExternalForce(frame.parentId, Frame::WORLD_FRAME, force, Frame::BODY_FRAME, frame.position);
+    Vec<3> force_b; force_b = frame.orientation * force;
+    setExternalForce(frame.parentId, Frame::BODY_FRAME, force_b, Frame::BODY_FRAME, frame.position);
   }
 
   /**
@@ -1015,7 +1073,7 @@ class ArticulatedSystem : public Object {
 
   /**
    * @return control mode. Can be either ControlMode::FORCE_AND_TORQUE or ControlMode::PD_PLUS_FEEDFORWARD_TORQUE */
-  ControlMode::Type getControlMode() const { return controlMode_; }
+  [[nodiscard]] ControlMode::Type getControlMode() const { return controlMode_; }
 
   /**
    * set PD targets.
@@ -1176,21 +1234,21 @@ class ArticulatedSystem : public Object {
   /**
    * Currently only supports "DO_NOT_COLLIDE_WITH_PARENT"
    * @return articulated system option */
-  ArticulatedSystemOption getOptions() const { return options_; }
+  [[nodiscard]] ArticulatedSystemOption getOptions() const { return options_; }
 
   /**
    * @return a vector of body names (following the joint order) */
-  const std::vector<std::string> &getBodyNames() const { return bodyName; }
+  [[nodiscard]] const std::vector<std::string> &getBodyNames() const { return bodyName; }
 
   /**
    * @return a vector of visualized bodies */
   std::vector<VisObject> &getVisOb() { return visObj; };
-  const std::vector<VisObject> &getVisOb() const { return visObj; };
+  [[nodiscard]] const std::vector<VisObject> &getVisOb() const { return visObj; };
 
   /**
    * @return a vector of visualized collision bodies */
   std::vector<VisObject> &getVisColOb() { return visColObj; };
-  const std::vector<VisObject> &getVisColOb() const { return visColObj; };
+  [[nodiscard]] const std::vector<VisObject> &getVisColOb() const { return visColObj; };
 
   /**
    * @param[in] visObjIdx visual object index. Following the order specified by the vector getVisOb()
@@ -1210,23 +1268,23 @@ class ArticulatedSystem : public Object {
 
   /**
    * @return the resource directory (for mesh files, textures, etc) */
-  const std::string &getResourceDir() const { return resourceDir_; }
+  [[nodiscard]] const std::string &getResourceDir() const { return resourceDir_; }
 
   /**
    * @return the resource directory (for mesh files, textures, etc) */
-  const std::string &getRobotDescriptionfFileName() const { return robotDefFileName_; }
+  [[nodiscard]] const std::string &getRobotDescriptionfFileName() const { return robotDefFileName_; }
 
   /**
    * @return the name of the URDF file (returns empty string if the robot was not specified by a URDF file) */
-  const std::string &getRobotDescriptionfTopDirName() const { return robotDefFileUpperDir_; }
+  [[nodiscard]] const std::string &getRobotDescriptionfTopDirName() const { return robotDefFileUpperDir_; }
 
   /**
    * @return the full path to the URDF file (returns empty string if the robot was not specified by a URDF file) */
-  const std::string &getRobotDescriptionFullPath() const { return fullURDFPath_; }
+  [[nodiscard]] const std::string &getRobotDescriptionFullPath() const { return fullURDFPath_; }
 
   /**
    * @return if the object was instantiated with raw URDF string, it returns the string */
-  const std::string &getRobotDescription() const { return robotDef_; }
+  [[nodiscard]] const std::string &getRobotDescription() const { return robotDef_; }
 
   /**
    * if the object was instantiated with raw URDF string, it exports the robot description to an URDF file
@@ -1274,10 +1332,26 @@ class ArticulatedSystem : public Object {
    * set the base position using an eigen vector
    * @param[in] pos position of the base */
   void setBasePos(const Vec<3> &pos);
+
   /**
-   * set the base orientation using an eigen vector
+   * set the base orientation using a raisim rotation matrix (i.e., Mat<3,3>)
    * @param[in] rot orientation of the base */
   void setBaseOrientation(const Mat<3, 3> &rot);
+
+  /**
+   * set the base orientation using a raisim quaternion (i.e., Vec<4>)
+   * @param[in] rot orientation of the base */
+    void setBaseOrientation(const Vec<4> &quat);
+
+  /**
+   * set the base position using an eigen vector
+   * @param[in] vel the linear velocity of the base */
+  void setBaseVelocity(const Vec<3> &vel);
+
+  /**
+   * set the base position using an eigen vector
+   * @param[in] vel the angular velocity of the base */
+  void setBaseAngularVelocity(const Vec<3> &vel);
 
   /**
    * set limits in actuation force. It can be also specified in the URDF file
@@ -1290,11 +1364,11 @@ class ArticulatedSystem : public Object {
 
   /**
    * @return the upper joint torque/force limit*/
-  const VecDyn &getActuationUpperLimits() const { return tauUpper_; }
+  [[nodiscard]] const VecDyn &getActuationUpperLimits() const { return tauUpper_; }
 
   /**
    * @return the lower joint torque/force limit*/
-  const VecDyn &getActuationLowerLimits() const { return tauLower_; }
+  [[nodiscard]] const VecDyn &getActuationLowerLimits() const { return tauLower_; }
 
   /**
    * change collision geom parameters.
@@ -1303,7 +1377,7 @@ class ArticulatedSystem : public Object {
    * For a sphere, {raidus}.
    * For a cylinder and a capsule, {radius, length}
    * For a box, {x-dim, y-dim, z-dim} */
-  void setCollisionObjectShapeParameters(size_t id, const std::vector<double> &params);
+  void setCollisionObjectShapeParameters(size_t id, const Vec<4> &params);
 
   /**
    * change collision geom offset from the joint position.
@@ -1331,7 +1405,7 @@ class ArticulatedSystem : public Object {
    * is not exactly equivalent in dynamics (due to gyroscopic effect). but it is a commonly used approximation.
    * It can also be expressed in the URDF file
    * @return the rotor inertia */
-  const VecDyn &getRotorInertia() const {
+  [[nodiscard]] const VecDyn &getRotorInertia() const {
     return rotorInertia_;
   }
 
@@ -1341,11 +1415,11 @@ class ArticulatedSystem : public Object {
    * @param[in] jointIndex the joint index
    * @return the joint type */
   Joint::Type getJointType(size_t jointIndex) { return jointType[jointIndex]; }
-  const Joint::Type getJointType(size_t jointIndex) const { return jointType[jointIndex]; }
+  [[nodiscard]] Joint::Type getJointType(size_t jointIndex) const { return jointType[jointIndex]; }
 
   /**
    * @return the number of joints (same as the number of bodies) */
-  size_t getNumberOfJoints() const { return nbody; }
+  [[nodiscard]] size_t getNumberOfJoints() const { return nbody; }
 
   /**
    * returns reference object of the joint
@@ -1366,27 +1440,27 @@ class ArticulatedSystem : public Object {
 
   /**
    * @return a mapping that converts body index to gv index */
-  const std::vector<size_t> &getMappingFromBodyIndexToGeneralizedVelocityIndex() const {
+  [[nodiscard]] const std::vector<size_t> &getMappingFromBodyIndexToGeneralizedVelocityIndex() const {
     return bodyIdx2GvIdx;
   }
 
   /**
    * @return a mapping that converts body index to gv index */
-  const std::vector<size_t> &getMappingFromBodyIndexToGeneralizedCoordinateIndex() const {
+  [[nodiscard]] const std::vector<size_t> &getMappingFromBodyIndexToGeneralizedCoordinateIndex() const {
     return bodyIdx2GcIdx;
   }
 
   /**
    * @return the object type (ARTICULATED_SYSTEM) */
-  ObjectType getObjectType() const final { return ARTICULATED_SYSTEM; }
+  [[nodiscard]] ObjectType getObjectType() const final { return ARTICULATED_SYSTEM; }
 
   /**
    * @return the body type (STATIC, KINEMATIC, or DYNAMIC) of the specified body. It is always DYNAMIC except for the fixed base*/
-  BodyType getBodyType(size_t bodyIdx) const final;
+  [[nodiscard]] BodyType getBodyType(size_t bodyIdx) const final;
 
   /**
    * @return the body type (STATIC, KINEMATIC, or DYNAMIC). It is always dynamic */
-  BodyType getBodyType() const final { return BodyType::DYNAMIC; };
+  [[nodiscard]] BodyType getBodyType() const final { return BodyType::DYNAMIC; };
 
   /**
    * @param[in] scheme the integration scheme. Can be either TRAPEZOID, SEMI_IMPLICIT, EULER, or RUNGE_KUTTA_4. We recommend TRAPEZOID for systems with many collisions.
@@ -1432,7 +1506,7 @@ class ArticulatedSystem : public Object {
 
   /**
    * Clears all external forces and torques */
-  void clearExternalForcesAndTorques() {
+  void clearExternalForcesAndTorques() final {
     isExternalForces_.resize(0);
     externalForceAndTorque_.resize(0);
     externalForceAndTorquePos_.resize(0);
@@ -1450,17 +1524,65 @@ class ArticulatedSystem : public Object {
   /**
    * @return springs Existing spring elements on joints */
   std::vector<SpringElement> &getSprings() { return springs_; }
-  const std::vector<SpringElement> &getSprings() const { return springs_; }
+  [[nodiscard]] const std::vector<SpringElement> &getSprings() const { return springs_; }
 
   /**
    *
    * @return parent parent[i] is a parent body id of the i^th body
    */
-  const std::vector<size_t>& getParentVector() const { return parent; }
+  [[nodiscard]] const std::vector<size_t>& getParentVector() const { return parent; }
+
+  /**
+   * @return sensor of the specified type
+   */
+  template<class T>
+  T* getSensor(const std::string& name) {
+    RSFATAL_IF(sensors_.find(name) == sensors_.end(), "Cannot find \""<<name<<"\"")
+    auto sensor = sensors_.find(name)->second;
+    RSFATAL_IF(sensor->getType() != T::getType(), "Type mismatch. "
+        << name << " has a type of " << toString(sensor->getType()) << " and the requested type is "
+        << toString(T::getType()))
+    return reinterpret_cast<T*>(sensor.get());
+  }
+
+  /**
+   * @return sensors on the robot
+   */
+  [[nodiscard]] std::unordered_map<std::string, std::shared_ptr<Sensor>>& getSensors() { return sensors_; }
+  [[nodiscard]] const std::unordered_map<std::string, std::shared_ptr<Sensor>>& getSensors() const { return sensors_; }
 
   // not recommended for users. only for developers
   void addConstraints(const std::vector<PinConstraintDefinition>& pinDef);
   void initializeConstraints();
+
+  /**
+   * If it is true, it also computes inverse dynamics and you can get the following properties:
+   * the force and torque (including the constrained directions) at the joint and acceleration of a body (necessary for IMU computation)
+   * @param flag
+   */
+  void setComputeInverseDynamics(bool flag) { computeInverseDynamics_ = flag; }
+
+  /**
+   *
+   * @return return if the inverse dynamics is computed
+   */
+  bool getComputeInverseDynamics() const { return computeInverseDynamics_; }
+
+  /**
+   * YOU MUST CALL raisim::ArticulatedSystem::setComputeInverseDynamics(true) before calling this method.
+   * This method returns the force at the specified joint
+   * @param jointId[in] the joint id
+   * @return force at the joint
+   */
+  const raisim::Vec<3>& getForceAtJointInWorldFrame(size_t jointId) const { return forceAtJoint_W[jointId]; }
+
+  /**
+   * YOU MUST CALL raisim::ArticulatedSystem::setComputeInverseDynamics(true) before calling this method.
+   * This method returns the torque at the specified joint
+   * @param jointId[in] the joint id
+   * @return torque at the joint
+   */
+  const raisim::Vec<3>& getTorqueAtJointInWorldFrame(size_t jointId) const { return torqueAtJoint_W[jointId]; }
 
  protected:
 
@@ -1478,8 +1600,6 @@ class ArticulatedSystem : public Object {
 
   void computeMassMatrix(MatDyn &M) noexcept;
 
-  void computeNonlinearities(const Vec<3> &gravity, VecDyn &b);
-
   void destroyCollisionBodies(dSpaceID id) final;
 
   /* This computes Delassus matrix necessary for contact force computation */
@@ -1487,21 +1607,35 @@ class ArticulatedSystem : public Object {
   void preContactSolverUpdate2(const Vec<3> &gravity, double dt, contact::ContactProblems& problems) final;
   void integrateWithoutContact(const Vec<3> &gravity, double dt);
 
-  void integrate(double dt, const Vec<3> &gravity) final;
+  void integrate(double dt, const World* world) final;
 
   void addContactPointVel(size_t pointId, Vec<3> &vel) final;
-
   void subContactPointVel(size_t pointId, Vec<3> &vel) final;
-
+  void addContactPointVel2(size_t pointId, Vec<3> &vel) final;
+  void subContactPointVel2(size_t pointId, Vec<3> &vel) final;
   void updateGenVelWithImpulse(size_t pointId, const Vec<3> &imp) final;
 
   void updateTimeStep(double dt) final;
 
   void updateTimeStepIfNecessary(double dt) final;
 
+  void updateSensorsIfNecessary(const raisim::World* world) final;
+
   inline void jacoSub(const raisim::SparseJacobian &jaco1, raisim::SparseJacobian &jaco, bool isFloatingBase);
 
   void setConstraintForce(size_t bodyIdx, const Vec<3> &pos, const Vec<3> &force) final;
+
+  void updateInertialMeasurementUnit(InertialMeasurementUnit* imu, Eigen::Vector3d& linAcc, Eigen::Vector3d& angVel);
+
+  /**
+   * Recursive Newton Euler algorithm computes inverse dynamics of the system.
+   * In other words, it computes the generalized force that results in the given given generalized acceleration.
+   * After calling this method, you can also call below methods
+   * @param gravity[in] The gravitational acceleration of the world (raisim::World::getGravity())
+   * @param ga[in] The generalized acceleration
+   * @param b[out] The generalized force
+   */
+  void recursiveNewtonEuler(const Vec<3> &gravity, bool includeExternalForces, bool updateJointForces, const VecDyn& ga, VecDyn &b);
 
  private:
 
@@ -1555,8 +1689,8 @@ class ArticulatedSystem : public Object {
   std::vector<raisim::Vec<3>> bodyLinVel_pc_W;
   std::vector<raisim::Vec<3>> bodyLinearAcc;
   std::vector<raisim::Vec<3>> bodyAngAcc;
-  std::vector<raisim::Vec<3>> propagatingForceAtJoint_W;
-  std::vector<raisim::Vec<3>> propagatingTorqueAtJoint_W;
+  std::vector<raisim::Vec<3>> forceAtJoint_W;
+  std::vector<raisim::Vec<3>> torqueAtJoint_W;
   std::vector<raisim::Mat<3, 3>> inertiaAboutJoint_W;
   std::vector<raisim::Mat<3, 3>> inertia_comW;
   std::vector<raisim::Vec<3>> sketch;
@@ -1588,14 +1722,14 @@ class ArticulatedSystem : public Object {
   std::vector<size_t> lambda;
 
   /// State variables
-  VecDyn gc_, gcOld_, gv_, udot_, gvOld_, gvAvg_, h_, gvERP_, gvTemp_, gvPreimpact_, gcRK_[4], gvRK_[4], slopeRK_[4];
+  VecDyn gc_, gcOld_, gv_, ga_, udot_, gvOld_, gvAvg_, h_, gvERP_, gvTemp_, gvPreimpact_, gcRK_[4], gvRK_[4], slopeRK_[4];
   MatDyn M_, Minv_, lT_;
 
   /// Contact variables
   std::vector<MatDyn> MinvJT_T;
   std::vector<VecDyn> j_MinvJT_T1D;
   VecDyn tauStar_, tau_, tauFF_;
-  VecDyn tauUpper_, tauLower_; // bounds
+  VecDyn tauUpper_, tauLower_, velLimits_; // bounds
   std::vector<size_t> bodyIdx2GvIdx, bodyIdx2GcIdx;
 
   std::vector<SparseJacobian> J_;
@@ -1616,6 +1750,7 @@ class ArticulatedSystem : public Object {
  private:
   size_t nbody, dof = 0, gcDim = 0;
   bool dampedDiagonalTermUpdated_ = false;
+  bool computeInverseDynamics_ = false;
   ControlMode::Type controlMode_ = ControlMode::FORCE_AND_TORQUE;
   raisim::SparseJacobian tempJaco_;
   double dt_ = 0.001;
@@ -1651,7 +1786,7 @@ class ArticulatedSystem : public Object {
   std::vector<PinConstraintDefinition> pinDef_;
 
   /// ABA
-  Mat<6, 6> MaInv_base;
+  Mat<6,6> MaInv_base;
   std::vector<Eigen::Matrix<double, 3, 3>, AlignedAllocator<Eigen::Matrix<double, 3, 3>, 32>> joint2Com_w_Skew;
 
   struct AbaData {
@@ -1683,6 +1818,9 @@ class ArticulatedSystem : public Object {
 
   std::vector<AbaData, AlignedAllocator<AbaData, 32>> ad_;
   std::vector<AbaData3, AlignedAllocator<AbaData3, 32>> ad3_;
+
+  /// sensors
+  std::unordered_map<std::string, std::shared_ptr<Sensor>> sensors_;
 };
 }